pyproject.toml
[tool.poetry]
name = "biometric-system-management-isp"
version = "0.1.0"
description = ""
authors = ["Andrew Chuang <andrewjschuang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
numpy = "^1.26.4"
pillow = "^10.2.0"
openpyxl = "^3.1.2"
face-recognition = "^1.3.0"
requests = "^2.31.0"
opencv-python = "^4.9.0.80"
websockets = "^12.0"
face-recognition-models = "^0.3.0"
setuptools = "^69.1.1"
flask-cors = "^4.0.0"
flask-socketio = "^5.3.6"
pymongo = "^4.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
